feat(StoryBook): add control attributes
